Move generated ruby client test to java.io.swagger.codegen.ruby package.
